# openapi/asoe2.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SeedCaseAttachmentRequest": {
        "additionalProperties": false,
        "description": "Sandbox attachment ingest. ``content_b64`` is optional \u2014 when omitted a\ndeterministic sample blob is stored so the download path can be exercised\nwithout supplying bytes.",
        "properties": {
          "content_b64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content B64"
          },
          "mime_type": {
            "default": "application/pdf",
            "title": "Mime Type",
            "type": "string"
          },
          "name": {
            "default": "sample.pdf",
            "title": "Name",
            "type": "string"
          }
        },
        "title": "SeedCaseAttachmentRequest",
        "type": "object"
      },
      "SeedEmailAttachmentAnchorsRequest": {
        "additionalProperties": false,
        "properties": {
          "anchors": {
            "items": {
              "$ref": "#/components/schemas/AnchorSpec"
            },
            "title": "Anchors",
            "type": "array"
          },
          "attachment_mime": {
            "title": "Attachment Mime",
            "type": "string"
          },
          "attachment_name": {
            "title": "Attachment Name",
            "type": "string"
          },
          "document_text": {
            "title": "Document Text",
            "type": "string"
          },
          "from_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Address"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          }
        },
        "required": [
          "document_text",
          "attachment_name",
          "attachment_mime",
          "anchors"
        ],
        "title": "SeedEmailAttachmentAnchorsRequest",
        "type": "object"
      },
      "SeedFinancialImpactRequest": {
        "additionalProperties": false,
        "properties": {
          "exception_id": {
            "title": "Exception Id",
            "type": "string"
          },
          "financial_impact_usd": {
            "title": "Financial Impact Usd",
            "type": "number"
          }
        },
        "required": [
          "exception_id",
          "financial_impact_usd"
        ],
        "title": "SeedFinancialImpactRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Deterministic, compliance-aware exception management API.",
    "title": "ASOE \u2014 Agentic System for Order-to-Cash Exceptions",
    "version": "0.3.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/_sandbox/cases/{case_id}/attachments": {
      "post": {
        "description": "Dev/demo producer for the attachment store (stand-in for the future\nemail-intelligence-agent / ADR-036 ingestion).\n\nPersists an attachment's bytes under (tenant, case_id) so the production\nread path \u2014 GET /cases/{case_id}/attachments/{attachment_id} \u2014 can be\nexercised end-to-end. Sandbox-only; real ingestion lands with ADR-036.",
        "operationId": "seed_case_attachment_api_v1__sandbox_cases__case_id__attachments_post",
        "parameters": [
          {
            "in": "path",
            "name": "case_id",
            "required": true,
            "schema": {
              "title": "Case Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedCaseAttachmentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Seed Case Attachment Api V1  Sandbox Cases  Case Id  Attachments Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Seed Case Attachment",
        "tags": [
          "sandbox"
        ]
      }
    },
    "/api/v1/_sandbox/seed/email-attachment-anchors": {
      "post": {
        "description": "Seed an EMAIL_ENTRY case with a stored attachment whose bytes contain\n``document_text`` and projected text-derived EvidenceAnchors, so the\ncomposer (`adapt_email_source` \u2192 `build_evidence_anchors`) emits located\nanchors with no new business logic (ADR-043 P1.1).",
        "operationId": "seed_email_attachment_anchors_api_v1__sandbox_seed_email_attachment_anchors_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedEmailAttachmentAnchorsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Seed Email Attachment Anchors Api V1  Sandbox Seed Email Attachment Anchors Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Seed Email Attachment Anchors",
        "tags": [
          "sandbox"
        ]
      }
    },
    "/api/v1/_sandbox/seed/financial-impact": {
      "post": {
        "description": "Stamp financial_impact_usd onto an exception's resolution_data.\n\nUsed by the Playwright four-eyes spec to push an existing record\npast HIGH_VALUE_OVERRIDE_THRESHOLD_USD so the next /disposition\ncall stages to PENDING_COSIGN. No new audit event is emitted \u2014\nthis is test fixture wiring, not a SOX-relevant state transition.",
        "operationId": "seed_financial_impact_api_v1__sandbox_seed_financial_impact_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedFinancialImpactRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Seed Financial Impact Api V1  Sandbox Seed Financial Impact Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Seed Financial Impact",
        "tags": [
          "sandbox"
        ]
      }
    },
    "/api/v1/_sandbox/seed/manual-order-intake": {
      "post": {
        "description": "Producer-side stand-in for the email-intelligence-agent.\n\nEmits a MANUAL_ORDER_INTAKE event by default and routes it through\nthe normal /exceptions/resolve graph so the dashboard's\n`event_type_received_total` gauge sees the canonical name.\nASOE_EMIT_LEGACY_EVENT_TYPES=1 flips emission to the transitional\nEMAIL_ORDER_ENTRY_REQUEST literal for dual-acceptance soaks.\n\nReturns the resulting exception_id, the event_type that was actually\nemitted, and whether the legacy flag was honoured \u2014 the caller's\nbackward-compat suite asserts on all three.",
        "operationId": "seed_manual_order_intake_api_v1__sandbox_seed_manual_order_intake_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedManualOrderIntakeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Seed Manual Order Intake Api V1  Sandbox Seed Manual Order Intake Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Seed Manual Order Intake",
        "tags": [
          "sandbox"
        ]
      }
    },
    "/api/v1/_sandbox/tenant/reset": {
      "post": {
        "description": "Clear in-memory exception records for a tenant.\n\nFor the in-memory store (the default sandbox mode), this wipes all\nexceptions belonging to the target tenant_id so the next spec starts\nfrom a clean slate. For the DB-backed store, this is a no-op \u2014 tests\nagainst a real DB spin up their own SQLite :memory: per run.",
        "operationId": "reset_tenant_api_v1__sandbox_tenant_reset_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Reset Tenant Api V1  Sandbox Tenant Reset Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reset Tenant",
        "tags": [
          "sandbox"
        ]
      }
    },
...
  }
}